Windows build fixes

+++ release/SciVisToolKit-11.02/MPIRawArrayIO.hxx
@@ -175,7 +175,7 @@ void CreateCartesianView(
   int decompStart[3];
   decomp.GetStartIndex(decompStart,domainStart);
 
-  unsigned long long nCells=decomp.Size();
+  unsigned long nCells=decomp.Size();
 
   // use a contiguous type when possible.
   if (domain==decomp)
